requirements.txt
semgrep>=1.77.0
pandas>=2.2.2
matplotlib>=3.9.0
streamlit>=1.37.0
langchain>=0.2.11
openai>=1.43.0
google-generativeai>=0.7.2
langchain-google-genai>=1.0.7
bcrypt>=4.0.1
fpdf2>=2.7.9
networkx>=3.3
